backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
pydantic>=2.0.0
httpx>=0.24.0
python-telegram-bot>=20.0
